Extract tool_call_execution.rs from turn.rs (parent #680)
Move the per-tool-call execution dispatch out of `turn.rs` into a
focused sibling module so `turn.rs` keeps shrinking toward
responsibility-focused units (parent meta-issue #680).

One production helper + 7 private helpers extracted as free functions
taking `&mut Agent` / `&Agent` (matching `actor_loop_flow` /
`reply_retry` / earlier vertical-slice precedent):

- `execute_tool_call` (production chokepoint: answer-only policy gate
  → empty-workspace scaffold policy gate → effective-tool-policy gate
  → cancel-flag check → Bash vs. non-Bash dispatch)
- `effective_tool_policy_error_for_execution` (private)
- `handle_tool_execution_rejection` (private: records working-memory
  + appends artifact-completion attempt for the rejection class)
- `tool_context` (private: builds `ToolContext` snapshot)
- `execute_bash_tool_call` (private: Bash execution + feedback +
  evidence + DangerousBlock unsafe-feedback path)
- `capture_pre_tool_hash_if_needed` (private: Write/Edit pre-tool hash)
- `execute_non_bash_tool_call` (private: non-Bash execution +
  touched-file + repo-edit evidence + Edit-failure feedback)
- `observe_evidence_from_bash_outcome` (private: Issue #606 T-1.6
  post-hoc Bash exit-0 → VerifierExitZero evidence +
  `last_verifier_invocation` record)

Behavior unchanged: same policy gate order, same `ToolContext` shape
(including `interactive_approval = io::stdin().is_terminal()`), same
Bash / non-Bash dispatch split, same feedback frame recording, same
Issue #606 T-1.6 BuildTest exit-0 evidence path + Issue #608 α-2
`last_verifier_invocation` recording invariants.

`answer_only_policy_error` and `effective_tool_policy_error` were
promoted from private to `pub(super)` so the sibling module can call
them as methods. Both stay in `turn.rs` for now; they're candidates
for a future policy-evaluation extraction.

`pub(super)` limited / no facade re-export (DR3-001). Callers updated
in `scaffold_pipeline.rs` (1 site) and `actor_loop_flow.rs` (4 sites)
to the free-fn form.

LOC delta:
- turn.rs: 1,387 → 1,076 (-311)
- new module: +360

3,114 lib tests pass; `cargo fmt` / `cargo clippy --lib --all-targets
-- -D warnings` clean.
